feat(lnch): static pages wear the app's chrome; footer gains highlight + GitHub
Owner feedback on the launch round. The authed header moves out of the
_authed layout into a shared AppHeader so the rules/terms/privacy pages
can render it for a signed-in member — the way back into the app is now
the same one every page has — while visitors get a slim bar routing to
the splash and sign-in (pre-claim sessions get the visitor bar too: the
full header's nav points into the subtree the username gate would
bounce them out of). The legal footer highlights the page the reader is
on (aria-current + emphasis), links the GitHub repo, and the privacy
contact becomes picksleagues@gmail.com.
